move data classes from one file to multiple files for better readability
