+++ mobile_app/lib/screens/inventory_details_view.dart
@@ -1,5 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:mobile_app/db/data_models.dart';
+import 'package:mobile_app/db/data_models/inventory.dart';
+import 'package:mobile_app/db/data_models/item_type.dart';
 import 'package:mobile_app/db/db_handler.dart';
 
 class InventoryDetailsView extends StatefulWidget {
