+++ mobile_app/lib/screens/main_views/inventory_list_view.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:mobile_app/db/data_models.dart';
+import 'package:mobile_app/db/data_models/inventory.dart';
 import 'package:mobile_app/db/db_handler.dart';
 // import 'package:mobile_app/db/item.dart';
 import 'package:mobile_app/widgets/add_inventory.dart';
