+++ mobile_app/lib/screens/main_views/new_item_screen.dart
@@ -1,5 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:mobile_app/db/data_models.dart';
+import 'package:mobile_app/db/data_models/item.dart';
+import 'package:mobile_app/db/data_models/item_type.dart';
 import 'package:mobile_app/db/db_handler.dart';
 import 'package:mobile_app/utils/item_arguments.dart';
 import 'package:mobile_app/widgets/add_items/found_item.dart';
