+++ mobile_app/lib/widgets/add_items/item_form.dart
@@ -1,5 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:mobile_app/db/data_models.dart';
+import 'package:mobile_app/db/data_models/inventory.dart';
+import 'package:mobile_app/db/data_models/item_type.dart';
 import 'package:mobile_app/db/db_handler.dart';
 
 class ItemForm extends StatefulWidget {
